[EH][Perf] Update perf documentation and tests
Update the perf tests README for Event Hubs as it was giving outdated
information. Also did some minor refactoring to the tests.
