+++ sdk/eventhub/azure-eventhub/tests/perfstress_tests/uamqp_event.py
@@ -2,13 +2,6 @@
 # Copyright (c) Microsoft Corporation. All rights reserved.
 # Licensed under the MIT License. See License.txt in the project root for license information.
 # --------------------------------------------------------------------------------------------
-
-import threading
-from concurrent.futures import ThreadPoolExecutor, as_completed
-import asyncio
-import time
-import os
-
 from azure.core.utils import parse_connection_string
 from uamqp.client import ReceiveClient
 from uamqp.async_ops.client_async import ReceiveClientAsync
@@ -61,7 +54,7 @@ class UamqpReceiveEventTest(EventPerfTest):
         config['consumer_group'] = "$Default"
         config['partition'] = "0"
         return config
-        
+
     def process_event_sync(self, message):
         try:
             message.annotations
